% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSimulation.R
\name{generateInitialGeneration}
\alias{generateInitialGeneration}
\title{Generate initial generation}
\usage{
generateInitialGeneration(numGenomes, numInputs, numOutputs)
}
\arguments{
\item{numGenomes}{Number of genomes}

\item{numInputs}{Number of inputs}

\item{numOutputs}{Number of outputs}
}
\description{
This function creates initial pool of genomes.
}
