% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperneat.R
\name{HyperNEATSimulation}
\alias{HyperNEATSimulation}
\title{Runs HyperNEAT simulation}
\usage{
HyperNEATSimulation(simulationParameters = NULL,
  substrateConfiguration = NULL)
}
\arguments{
\item{simulationParameters}{List of parameters that control the simulation}

\item{substrateConfiguration}{List of substrate configurations}
}
\description{
This function runs the HyperNEAT simulation according to the problem configuration
}
