% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSimulation.R
\name{generateInitialGeneration}
\alias{generateInitialGeneration}
\title{Generate initial generation}
\usage{
generateInitialGeneration(numGenomes, numInputs, numOutputs)
}
\arguments{
\item{numGenomes}{Number of individual genomes}

\item{numInputs}{Number of inputs}

\item{numOutputs}{Number of outputs}
}
\value{
Returns a list of length \code{numGenomes} of individual genomes with all input nodes connected to all
output nodes. Weights are initialised with runif(n=1,min=0,max=1) and activation functions are rectified
linear units.
}
\description{
This function creates initial pool of genomes.
}
